Moved managed instance queries into a separate file
